# qlife-latest/qlife-server/openapi.yaml
openapi: 3.0.3
info:
  title: QLife API
  description: |
    REST API for the QLife mental-health platform. Supports three primary roles:
    - **USER** – a registered end-user seeking mental-health support
    - **PROFESSIONAL** – a verified mental-health professional
    - **ADMIN** – a platform administrator

    All protected endpoints require a Bearer JWT obtained from the Cognito-backed
    authentication flow. The token is passed in the `Authorization` header as
    `Bearer <token>`.

    All successful responses wrap their payload in a top-level `data` key.
    Error responses use `{ "error": { "code": "...", "message": "..." } }`.
  version: 1.0.0
  contact:
    name: QLife Engineering
servers:
  - url: https://api.qlife.app/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /appointments:
    get:
      tags: [Appointments]
      operationId: listAppointments
      summary: List appointments
      description: |
        Returns a paginated list of appointments relevant to the authenticated
        account. USERs see their own appointment requests; PROFESSIONALs see
        appointment requests directed to them. Requires role **USER** or
        **PROFESSIONAL**.
      parameters:
        - $ref: '#/components/parameters/PageQuery'
      responses:
        '200':
          description: Appointment list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      appointments:
                        type: array
                        items:
                          $ref: '#/components/schemas/Appointment'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
              example:
                data:
                  appointments:
                    - id: "apt-0000-0000-0000-000000000001"
                      status: "PENDING"
                      requestedStartAt: "2025-07-15T09:00:00.000Z"
                      requestMessage: "I would like to discuss anxiety management."
                      profileShareGranted: true
                      createdAt: "2025-06-10T08:00:00.000Z"
                  pagination:
                    page: 1
                    pageSize: 20
                    total: 3
                    hasMore: false
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Appointments]
      operationId: requestAppointment
      summary: Request an appointment
      description: |
        Creates a new appointment request from the authenticated USER to a
        specific professional. Requires role **USER**.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestAppointmentRequest'
            example:
              professionalProfileId: "prof-0000-0000-0000-000000000001"
              requestedStartAt: "2025-07-15T09:00:00.000Z"
              requestMessage: "I would like to discuss anxiety management."
              profileShareGranted: true
      responses:
        '200':
          description: Appointment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      appointmentId:
                        type: string
                        format: uuid
              example:
                data:
                  appointmentId: "apt-0000-0000-0000-000000000001"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /appointments/{id}:
    get:
      tags: [Appointments]
      operationId: getAppointment
      summary: Get appointment detail
      description: |
        Returns the full detail of a single appointment. The authenticated
        account must be a party to the appointment (USER or PROFESSIONAL).
      parameters:
        - name: id
          in: path
          required: true
          description: Appointment UUID
          schema:
            type: string
            format: uuid
            example: "apt-0000-0000-0000-000000000001"
      responses:
        '200':
          description: Appointment detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      appointment:
                        $ref: '#/components/schemas/AppointmentDetail'
              example:
                data:
                  appointment:
                    id: "apt-0000-0000-0000-000000000001"
                    status: "ACCEPTED"
                    requestedStartAt: "2025-07-15T09:00:00.000Z"
                    scheduledStartAt: "2025-07-15T10:00:00.000Z"
                    requestMessage: "I would like to discuss anxiety management."
                    professionalMessage: "I look forward to our session."
                    meetingLink: "https://meet.example.com/session-abc"
                    profileShareGranted: true
                    seenByProfessional: true
                    createdAt: "2025-06-10T08:00:00.000Z"
                    updatedAt: "2025-06-11T09:00:00.000Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /appointments/{id}/seen:
    post:
      tags: [Appointments]
      operationId: markAppointmentSeen
      summary: Mark appointment as seen (Professional)
      description: |
        Marks an appointment request as seen by the professional. Requires role
        **PROFESSIONAL** and the appointment must belong to them.
      parameters:
        - name: id
          in: path
          required: true
          description: Appointment UUID
          schema:
            type: string
            format: uuid
            example: "apt-0000-0000-0000-000000000001"
      responses:
        '200':
          description: Updated appointment
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      appointment:
                        $ref: '#/components/schemas/AppointmentDetail'
              example:
                data:
                  appointment:
                    id: "apt-0000-0000-0000-000000000001"
                    status: "PENDING"
                    seenByProfessional: true
                    requestedStartAt: "2025-07-15T09:00:00.000Z"
                    createdAt: "2025-06-10T08:00:00.000Z"
                    updatedAt: "2025-06-12T07:30:00.000Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /appointments/{id}/respond:
    post:
      tags: [Appointments]
      operationId: respondToAppointment
      summary: Respond to an appointment (Professional)
      description: |
        Allows a PROFESSIONAL to accept, decline, or propose a reschedule for
        an appointment request. Requires role **PROFESSIONAL** and the
        appointment must belong to them.

        - `ACCEPTED`: confirm the appointment; provide `scheduledStartAt` and
          optionally `meetingLink`.
        - `DECLINED`: decline the request; optionally provide a
          `professionalMessage`.
        - `RESCHEDULE_PROPOSED`: propose a new time; provide `scheduledStartAt`.
      parameters:
        - name: id
          in: path
          required: true
          description: Appointment UUID
          schema:
            type: string
            format: uuid
            example: "apt-0000-0000-0000-000000000001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RespondAppointmentRequest'
            example:
              action: "ACCEPTED"
              scheduledStartAt: "2025-07-15T10:00:00.000Z"
              professionalMessage: "I look forward to our session."
              meetingLink: "https://meet.example.com/session-abc"
      responses:
        '200':
          description: Updated appointment after response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      appointment:
                        $ref: '#/components/schemas/AppointmentDetail'
              example:
                data:
                  appointment:
                    id: "apt-0000-0000-0000-000000000001"
                    status: "ACCEPTED"
                    scheduledStartAt: "2025-07-15T10:00:00.000Z"
                    meetingLink: "https://meet.example.com/session-abc"
                    professionalMessage: "I look forward to our session."
                    seenByProfessional: true
                    createdAt: "2025-06-10T08:00:00.000Z"
                    updatedAt: "2025-06-12T08:00:00.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------- Assessments ---------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by Amazon Cognito. Pass as Authorization: Bearer <token>.
  parameters:
    PageQuery:
      name: page
      in: query
      required: false
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "professionalProfileId must be a UUID"
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Unauthorized"
    Forbidden:
      description: Authenticated account lacks the required role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "FORBIDDEN"
              message: "Forbidden"
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error description
              example: "professionalProfileId must be a UUID"
    Pagination:
      type: object
      description: Pagination metadata included in list responses
      required: [page, pageSize, total, hasMore]
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        pageSize:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        hasMore:
          type: boolean
          example: true
...
    Appointment:
      type: object
      description: Summary appointment record for list views
      properties:
        id:
          type: string
          format: uuid
          example: "apt-0000-0000-0000-000000000001"
        status:
          type: string
          enum: [PENDING, ACCEPTED, DECLINED, RESCHEDULE_PROPOSED, CANCELLED, COMPLETED]
          example: "PENDING"
        requestedStartAt:
          type: string
          format: date-time
          example: "2025-07-15T09:00:00.000Z"
        requestMessage:
          type: string
          nullable: true
          example: "I would like to discuss anxiety management."
        profileShareGranted:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2025-06-10T08:00:00.000Z"
    AppointmentDetail:
      allOf:
        - $ref: '#/components/schemas/Appointment'
        - type: object
          properties:
            scheduledStartAt:
              type: string
              format: date-time
              nullable: true
              example: "2025-07-15T10:00:00.000Z"
            professionalMessage:
              type: string
              nullable: true
              example: "I look forward to our session."
            meetingLink:
              type: string
              nullable: true
              example: "https://meet.example.com/session-abc"
            seenByProfessional:
              type: boolean
              example: true
            updatedAt:
              type: string
              format: date-time
              example: "2025-06-11T09:00:00.000Z"
    RequestAppointmentRequest:
      type: object
      required: [professionalProfileId, requestedStartAt]
      properties:
        professionalProfileId:
          type: string
          format: uuid
          example: "prof-0000-0000-0000-000000000001"
        requestedStartAt:
          type: string
          format: date-time
          description: Requested appointment start time (ISO 8601)
          example: "2025-07-15T09:00:00.000Z"
        requestMessage:
          type: string
          maxLength: 1000
          example: "I would like to discuss anxiety management."
        profileShareGranted:
          type: boolean
          example: true
    RespondAppointmentRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [ACCEPTED, DECLINED, RESCHEDULE_PROPOSED]
          example: "ACCEPTED"
        scheduledStartAt:
          type: string
          format: date-time
          description: Confirmed or proposed new start time (required for ACCEPTED or RESCHEDULE_PROPOSED)
          example: "2025-07-15T10:00:00.000Z"
        professionalMessage:
          type: string
          maxLength: 1000
          example: "I look forward to our session."
        meetingLink:
          type: string
          maxLength: 2000
          example: "https://meet.example.com/session-abc"

    # ---- Assessments -------------------------------------------------------
...